ref: DeleteTrackPage: Separated DeleteTrack Form into a different component for better readability and usage
